code/requirements.txt
GPUtil==1.4.0
pytorch-transformers==1.2.0
torch>=1.10.2
faiss-cpu==1.7.2
